feat(portal): add page-level metadata to all 24 pages
Split 22 client component pages into page-client.tsx (client content)
and page.tsx (server wrapper with metadata export). Added metadata
directly to 2 existing server component pages (onboarding, admin/users).

Closes #169
